requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
discord-py-slash-command>=4.2.1

# OpenAI Integration
openai>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Scheduling
apscheduler>=3.10.0

# Image Processing
pillow>=10.0.0

# HTTP Client
aiohttp>=3.8.0

# Environment Management
python-dotenv>=1.0.0

# Timezone Support
pytz>=2023.3

# Utilities
python-dateutil>=2.8.2
asyncpg>=0.28.0

# Logging and Monitoring
structlog>=23.1.0